requirements.txt
black==23.7.0
clean-fid==0.1.35
isort==5.12.0
mypy==1.5.0
numpy==1.25.2
omegaconf==2.3.0
openTSNE==1.0.0
pandas==2.0.3
Pillow==10.0.0
pip==24.2
pytorch-lightning==2.0.6
PyYAML==6.0.1
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
tensorboard==2.14.0
torch==2.0.1
torch-ema==0.3
torchmetrics==1.0.3
torchvision==0.15.2
tqdm==4.66.1
multi-level-split @ git+https://github.com/lmkoch/multi-level-split